FunctionDepsFinder: Factor out closure logic
Factor-out the first part of the Closure computation so we can use it
in other parts of clang-extract.
